feat(web): Improve download page and image handling
- Replace local image with Supabase-hosted image
- Add Apple Silicon download route with redirect
- Update download button to link directly to download
- Optimize image loading with custom Image component
- Remove unnecessary separators in download page
- Enhance route configuration for better navigation
