+++ apps/web/src/routes/_view/download/apple-silicon.tsx
@@ -1,0 +1,9 @@
+import { createFileRoute, redirect } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/_view/download/apple-silicon")({
+  beforeLoad: async () => {
+    throw redirect({
+      href: "https://desktop.hyprnote.com/download/latest/dmg-aarch64?channel=stable",
+    });
+  },
+});
